+++ .storybook/main.js
@@ -3,14 +3,14 @@
  */
 const config = {
   stories: [
-    '../src/stories/Button.stories.tsx',
-    '../src/stories/Theme.stories.mdx',
-    '../src/stories/*.stories.@(js|jsx|ts|tsx)'
+    '../src/stories/*.stories.@(js|jsx|ts|tsx|mdx)',
+    '../src/stories/**/*.stories.@(js|jsx|ts|tsx|mdx)',
   ],
   addons: [
     '@storybook/addon-essentials',
     'storybook-addon-themes',
-    '@storybook/addon-a11y'
+    '@storybook/addon-a11y',
+    'storybook-addon-rtl'
   ],
   framework: '@storybook/react',
   features: {