+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 Project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [0.1.3] - 10-9-2022
+- adds lint-staged and Storybook tooling
+- refactors colors
+- adds `variant` and `loading` props and secondary button variant
+- updates light/dark themes
 
 ### [0.1.2] - 10-2-2022
 - Adds support for Tailwind CSS (#4)